cjs support
