+++ package.json
@@ -3,15 +3,17 @@
   "version": "1.0.4-alpha.3",
   "description": "Node.js library used to integrate with Descope",
   "typings": "./dist/index.d.ts",
-  "main": "./dist/index.js",
+  "main": "dist/cjs/index.cjs.js",
+  "module": "./dist/index.esm.js",
+  "types": "dist/index.d.ts",
   "exports": {
-    "require": "./dist/index.js",
-    "import": "./module.mjs"
+    "require": "./dist/cjs/index.cjs.js",
+    "import": "./dist/index.esm.js"
   },
+  "type": "module",
   "files": [
     "dist/**/*"
   ],
-  "type": "module",
   "repository": {
     "type": "git",
     "url": "git://github.com/descope/node-sdk.git"
@@ -20,14 +22,14 @@
     "node": ">= 14.0.0"
   },
   "scripts": {
-    "build": "rm -rf dist types && babel lib --out-dir dist --extensions '.ts' && tsc --declaration --outDir dist --emitDeclarationOnly",
+    "build": "rimraf dist && rollup -c",
     "quick": "npm i && npm run build && cd examples/es6/ && npm i && npm start",
     "format": "prettier --write lib",
     "check-format": "prettier --check lib",
     "size": "npm run build && size-limit",
     "lint": "eslint lib/**/*.ts --fix",
     "leaks": "bash ./scripts/gitleaks/gitleaks.sh",
-    "test": "jest --forceExit --config jest.config.ts --coverage --ci --detectOpenHandles .*test.*",
+    "test": "jest --forceExit --coverage --ci --detectOpenHandles .*test.*",
     "prepublishOnly": "npm run build"
   },
   "author": "Descope",
@@ -44,6 +46,11 @@
     "@babel/core": "^7.17.12",
     "@babel/preset-env": "^7.17.12",
     "@babel/preset-typescript": "^7.17.12",
+    "@rollup/plugin-commonjs": "^22.0.0",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^13.3.0",
+    "@rollup/plugin-replace": "^4.0.0",
+    "@rollup/plugin-typescript": "^8.3.0",
     "@size-limit/preset-small-lib": "^8.0.0",
     "@types/jest": "^28.0.0",
     "@types/jsonwebtoken": "^8.5.8",
@@ -69,6 +76,15 @@
     "jsdoc": "^3.6.10",
     "nock": "^13.2.4",
     "prettier": "2.6.2",
+    "rollup": "^2.62.0",
+    "rollup-plugin-auto-external": "^2.0.0",
+    "rollup-plugin-browsersync": "^1.3.3",
+    "rollup-plugin-delete": "^2.0.0",
+    "rollup-plugin-dts": "^4.2.2",
+    "rollup-plugin-esbuild": "^4.9.1",
+    "rollup-plugin-inject-process-env": "^1.3.1",
+    "rollup-plugin-livereload": "^2.0.5",
+    "rollup-plugin-terser": "^7.0.2",
     "ts-jest": "^28.0.3",
     "ts-node": "^10.8.2",
     "typescript": "^4.6.4"
@@ -83,7 +99,7 @@
     }
   ],
   "dependencies": {
-    "@descope/core-js-sdk": "0.0.41-alpha.7",
+    "@descope/core-js-sdk": "0.0.41-alpha.8",
     "jose": "4.8.1",
     "node-fetch": "2.6.7"
   }
